requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
wave>=0.0.2
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
sounddevice>=0.4.6
scipy>=1.11.0
pytesseract>=0.3.10
chromadb>=0.5.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai-whisper>=20231117
stripe>=5.0.0

# Flask backend for Wave 1 PR
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
werkzeug==3.0.1

# Production backend dependencies
PyJWT==2.8.0
bcrypt==4.1.2

# Wave 3 dependencies
chromadb==0.4.18
torch>=2.6.0
openai-whisper==20231117
transformers==4.35.2
numpy==1.24.4
librosa==0.10.1

# Wave 5 dependencies - Task/PR/CI-aware
websockets==12.0
PyYAML==6.0.1

# Fallback dependencies
sqlite3

# Optional advanced features
# whisperx  # Commented out due to Python 3.13 compatibility issues
pydub>=0.25.1
webrtcvad>=2.0.10

# Additional dependencies for advanced features
psutil>=5.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
PyPDF2>=3.0.0

# ========== ENHANCED SDLC FEATURES ==========

# Database and caching
redis>=4.5.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# JIRA and project management
jira>=3.5.0
atlassian-python-api>=3.40.0

# GitHub integration
PyGithub>=1.59.0
gitpython>=3.1.40

# Audio processing and speaker diarization
pyannote.audio>=3.1.0
speechbrain>=0.5.15

# Advanced AI/ML features
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Security and encryption
cryptography>=41.0.0
bcrypt>=4.0.0

# Monitoring and observability
prometheus-client>=0.17.0
structlog>=23.1.0

# Task scheduling and background jobs
celery>=5.3.0
redis-py-cluster>=2.1.3

# API documentation
flask-restx>=1.2.0
marshmallow>=3.20.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Mobile and cross-platform
fastapi>=0.103.0  # Alternative to Flask for mobile APIs
uvicorn>=0.23.0

# IDE integration helpers
watchdog>=3.0.0  # File system monitoring
psutil>=5.9.0   # Process monitoring
pdfplumber>=0.9.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Optional speaker diarization
pyannote.audio>=3.1.0
torch>=2.6.0
torchaudio>=2.6.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional dependencies for comprehensive AI Mentor implementation
gitpython>=3.1.40
markdown>=3.5.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
docker>=6.1.0
kubernetes>=28.1.0
boto3>=1.34.0
docx2txt>=0.8
opencv-python>=4.8.0
pytesseract>=0.3.10
sounddevice>=0.4.6
psutil>=5.9.0

# Machine Learning and AI dependencies
librosa>=0.10.1
transformers>=4.35.0
torch>=2.6.0
torchaudio>=2.6.0
scikit-learn>=1.3.0
wav2vec2>=0.1.0

# Additional integrations
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
exchangelib>=5.2.0
msal>=1.24.0

# Document processing (used by profile_manager)
PyPDF2>=3.0.0
docx2txt>=0.8

# Audio/video processing 
sounddevice>=0.4.6
psutil>=5.9.0